All circular dependencies removed
