+++ src/ducks/Studio/Template/TemplateContextMenu/Option.js
@@ -1,4 +1,6 @@
 import React from 'react'
+import cx from 'classnames'
+import Button from '@santiment-network/ui/Button'
 import styles from '../Dialog/LoadTemplate/Template.module.scss'
 
 const Option = props => (
