+++ src/ducks/Watchlists/Widgets/Table/AssetsTable.js
@@ -19,7 +19,8 @@ import AssetsToggleColumns from './AssetsToggleColumns'
 import Filter from '../Filter'
 import SaveAs from '../../Actions/SaveAs'
 import { useUserSubscriptionStatus } from '../../../../stores/user/subscriptions'
-import { COLUMNS, COMMON_SETTINGS, COLUMNS_SETTINGS } from './asset-columns'
+import { COLUMNS } from './asset-columns'
+import { COMMON_SETTINGS, COLUMNS_SETTINGS } from './columns'
 import ScreenerSignalDialog from '../../../Signals/ScreenerSignal/ScreenerSignalDialog'
 import { markedAsShowed } from '../../../SANCharts/SidecarExplanationTooltip'
 import { EXPLANATION_TOOLTIP_MARK } from '../../../Studio/Template/LayoutForAsset/LayoutForAsset'
