+++ src/pages/Account/gql.js
@@ -1,0 +1,11 @@
+import gql from 'graphql-tag'
+
+export const NEWSLETTER_SUBSCRIPTION_MUTATION = gql`
+  mutation changeNewsletterSubscription(
+    $subscription: NewsletterSubscriptionType
+  ) {
+    changeNewsletterSubscription(newsletterSubscription: $subscription) {
+      newsletterSubscription
+    }
+  }
+`
