+++ src/pages/feed/GeneralFeed/FeedList/FeedList.js
@@ -5,7 +5,7 @@ import FeedItemRenderer from '../FeedItemRenderer/FeedItemRenderer'
 import SonarFeedRecommendations from '../../../SonarFeed/SonarFeedRecommendations'
 import Loader from '@santiment-network/ui/Loader/Loader'
 import MakeProSubscriptionCard from '../MakeProSubscriptionCard/MakeProSubscriptionCard'
-import { addDays, getDateFormats } from '../../../../utils/dates'
+import { getDateFormats } from '../../../../utils/dates'
 import StoriesList from '../../../../components/Stories/StoriesList'
 import PageLoader from '../../../../components/Loader/PageLoader'
 import styles from './FeedList.module.scss'
