+++ src/pages/feed/GeneralFeed/FeedList/dates.js
@@ -1,2 +1,4 @@
+import { addDays } from '../../../../utils/dates'
+
 export const TODAY = new Date().toLocaleDateString()
 export const YESTERDAY = addDays(new Date(), -1).toLocaleDateString()
